moving translators to standard_to_enhanced folder, updating documentation
