+++ wzdx/standard_to_enhanced/cotrip_translator.py
