+++ wzdx/standard_to_enhanced/navjoy_translator.py
